Rodar todos os raspadores em um único processo
